server/services/adventure/pyproject.toml
[tool.poetry]
name = "adventure"
version = "0.1.0"
description = ""
authors = ["DucHoang-UIT <hoang.pham@executionlab.asia>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
django-debug-toolbar = "^4.3.0"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.3.0"}
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.1"
pillow = "^10.3.0"
django-cors-headers = "^4.3.1"
python-dotenv = "^1.0.1"
django-allauth = "^0.63.3"
requests = "^2.32.3"
requests-oauthlib = "^2.0.0"
python3-openid = "^3.2.0"
pyjwt = "^2.8.0"
python-keycloak = "^4.1.0"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
